+++ src/detection/terminalfont/terminalfont.c
@@ -160,7 +160,7 @@ static const char* detectFromWTImpl(FFstrbuf* content, FFstrbuf* name, double* s
     #include <shlobj.h>
 #endif
 
-static void detectFromWindowsTeriminal(const FFstrbuf* terminalExe, FFTerminalFontResult* terminalFont)
+static void detectFromWindowsTerminal(const FFstrbuf* terminalExe, FFTerminalFontResult* terminalFont)
 {
     //https://learn.microsoft.com/en-us/windows/terminal/install#settings-json-file
     FF_STRBUF_AUTO_DESTROY json = ffStrbufCreate();
@@ -174,7 +174,7 @@ static void detectFromWindowsTeriminal(const FFstrbuf* terminalExe, FFTerminalFo
         char* pathEnd = jsonPath + strlen(jsonPath);
         strncpy(pathEnd, ".portable", sizeof(jsonPath) - (size_t) (pathEnd - jsonPath) - 1);
 
-        if(ffPathExists(jsonPath, FF_PATHTYPE_FILE))
+        if(ffPathExists(jsonPath, FF_PATHTYPE_ANY))
         {
             strncpy(pathEnd, "settings\\settings.json", sizeof(jsonPath) - (size_t) (pathEnd - jsonPath) - 1);
             if(!ffAppendFileBuffer(jsonPath, &json))
@@ -432,7 +432,7 @@ static bool detectTerminalFontCommon(const FFTerminalResult* terminal, FFTermina
     //Used by both Linux (WSL) and Windows
     else if(ffStrbufIgnCaseEqualS(&terminal->processName, "Windows Terminal") ||
         ffStrbufIgnCaseEqualS(&terminal->processName, "WindowsTerminal.exe"))
-        detectFromWindowsTeriminal(&terminal->exe, terminalFont);
+        detectFromWindowsTerminal(&terminal->exe, terminalFont);
     #endif
 
     else